requirements.txt
fastapi==0.95.2
aiohttp==3.8.1
twilio==7.18.0
websockets==11.0.3
python-dotenv==0.21.0
uvicorn==0.22.0
